BASW-65: Fix Bug On Membership Creation
A bug was injected into membership creation by hook that altered membership
status, as it assumed the hook would only be called once membership was
created. This waas not the case, as the hook was being called prior to
creating the membership to calculate status.

Fixed by checking if membership ID exists, prior to altering calculated status
and only proceeding with alteration if ID is set.