BASW-90: Fix Membership Creaation and Renewal Forms Validation
Validation was not being executed for for empty installments and frequencies.
Renamed method that checked if a payment plan is being used so it is clear it
also validates if more than one contribution is being used to pay. Added
validation of number of installments. Added red *'s to show fields
'installments', 'installments_frequency', and 'installments_frequency_unit'
are required.