BASW-90: Fix Styling and Usability Problems
Validation Messages for installment quantity and frequency were breaking
styling of the form. Also, when validation failed, the form was being shown
again, but as if a single contribution was used, instead of a payment plan.

Altered template to remove messages that were causing break and added logic so
if validation fails, the state of the form is preserved according to what
payment option was chosen, contribution or payment plan.