BASW-118: Fix Issues on Recurring Contribution Update
When a contribution is set to auto-renew, CiviCRM makes it uneditable.
Implemented the links hook to re-enable edit permission for auto-renew manual
payment plans. Also, updated contrib_recur_id column if auto-renew is true.

Added logic to update date of next contribution on recurring contribution.

Fixed a problem where if auto-renew was set as false, the value wasn't changed
on the database.