BASW-91: Show if Payment Processors are Live or Not on Settings
Manually adding a new payment processor creates two different items, with the
same name, one with is_test = 0, the other one with is_test = 1. This made it
impossible to know which was which on the extension's settings form, where we
were just showing the processor's name.

Added 'Live - ' and 'Test - ' prefeixes correspondingly to each processor, so
users can know which is which.