BASW-602: Post Process Line Items to Add to Recurring Contribution
For recurring contributions used to pay for memberships we need to make a copy
of all line items associated to the first installment, in order to be able to
edit them later and propagate the changes among remaining pending installments
if need be.

Added post hook that pos processes line items, and checks if the associated
contribution is being used to pay for a membership. If so, a copy of the line
item is made an associated to the recurring contribution via the new
membershipextras_contribution_recur_line_item table.