BASW-218: Implement Removal of Line Items from Current Period
Implemented functionality to remove line items from current period by adding a
new form that confirms the line item is to be deleted and on processing changes
the amounts for related contributions and their line items to reflect the
change. Also cretes financial transactions for eache contribution with its
financial line items to reflect the amount reduction.