BASW-218: Fix Auto-Renew for Recurring Line Items
When recurring line items where being created, they were adopting the value
for auto-renew from the recurring contribution as it was set at the DB. The
problem is by this time, if the membership was set to auto-renew, the
recurring will always be set to false, as this value is stored as the last
step of the form's post-processing.

Fixed by adding  method that setsrecurring line items auto-renew to true at
the same time it is set for the recurring contribution.

Also fixed removal form so daate picker only appears if the checkbox is
ticked.

Changed getLineItems method so only line items that have not been removed
appear on the recurring contribution line item's list.