BASW-220: Fix Line Items Referencing Single Membership
Ran into bug/feature with CiviCRM on line item creation. All membership
line items are created associated to main membership in a price set before the
other memberships are created. Once the memberships are created, a query is
run to update the line items and associate them to the real memberships. This
was causing the recurring line items to be referencing only one membership,
even if visually they seemed ok. Fixed by running a similar query after the
other memberhips are created, when membership payments are recorded, which is
the way CiviCRM does it.