BASW-220: Check if Line Item Exists for Contributions
An error was being thrown when a line item for a membership was removed and
then added again, as there is a unique index on the line items table that does
not allow the same membership to be added for a contribution with the same
PriceFieldValue. Fixed by checking if line item exists and if so, updating the
amount.