BASW-263: Fix Bug Causing Incomplete Instalments on Renewal
When adding new line items to next period only, the 'entity_table' and
'entity_id' fields are set to point to the recurring conribution. On renewal,
price was being updated to latest when entity_table was different to
'civicrm_membership', causing line unit price, total and tax to be null,
causing exceptions to be thrown.

Refactored line item copying and membership minimum fee calulcation so the
case where the line item is for a membership that does not exist yet
calculates  the price from its price_value.

Enveloped the whole renewal of a payment plan within a transaction, to ensure
data is left in a consistent state if any exceptions are thrown on processing.