BASW-67: Fix Double Recurring Contribution Bug
When creating a membership and paying for it with a single instlment payment
plan, two recurring contributions were being created. This was because the
logic to determine if the membership was a payment plan was checking only if
instalments were exclusively more than one, creating  second recurring
contribution falsely thinking it was a renewal, instead of a cretion.

Fixed by changing logic to checko for instalments > 0.