BASW-122: Fix Problem When Completing Payment
When editing a pending contribuion for a payment plan with a single instalment
and that payment plan corresponded to a membership with an overridden status,
the system took a different path to update, including a method run when a
contribution is updated and its status is changed, that checked the recurring
contribution and changed its status to 'Completed' if all installments had
been paid. This is a problem on payment plans with only one installment, as
this is used as an auto-renewing contribution, and thus, should always be
'In Progress'.

Fixed problem by adding a pre hook on the Recurring Contribution entity, that
calculates the plan's status according to our rules.