ASE-297: Fix Creation of Payment Plans with Several Line Items
When a payment plan was creted using a price set, and some of the selected
items had tax while others didn't, a financial transaction for the tax was
still tried to be created for line items without tax, cusing an error when the
financial item was being created, as the 'amount' field was not included in
the insert query because the value was null for line items without tax.

Fixed by only creating financial item for tax if the line item has a tax
amount defined.