BASW-261: Show Recurring Line Items With End Date for Multi-Installment Plans
When a multiple installment payment plan is completed, the end date of the
recurring line items is also set. This was causing these line items to be
removed from the list of line items shown on the 'View/Edit Future
Installments', both from current and next periods.

Fixed by only filtering out line items with an end date if it is a payment
plan with 1 or less installments.