BASW-261: Fix Problem With Line Item Amounts
Line item amount for the new recurring contribution was being calculated as
the full amount for the membership, as the calculation for the amount per
installment was failing, as it required acces to the number of installments of
the new recurring contribution, but the value was not still set.

Fixed by obtaining the value from the current recurring contribution instead.