BASW-222: Implement Validation of Membership Type
A membership type could be added to next period, and then, added to current
period without auto-renew. If auto-renew is then set, it resulted line item to
be duplicated in next period, as when the auto-renew check-box was set, the
line item was added as a donation on next period. So each time auto-renew
check-box was se on, a new line item was added to next period.

Fixed by implementing validation of membership type, so if a line item for
that membership type already exists on next period, a waning message is shown
and no action is taken. Also, if validation passes ok, current line item is
just set to be auto-renewed and recurring contribution - membership
relationship is restored.