BASW-219: Fix Bug When Adding Line Items to Plans With no Contributions
When a membership line item was added not to auto-renew, and then was set to
auto-renew for a payment plan with 0 installments, the line item got deleted
from the database. This happened because although the line item was created,
since all contributions were in the past and/or completed, there were no
payments associated to the new membership.

Now, when setting the line item to auto-renew, the membership needs to be
updated to store the recurring contribution's ID in the contribution_recur_id
field of the Membership entity. On CiviCRM core, the create() method of the
membership's BAO checks if the membership has any payments. If it doesn't,
it assumes the membership type has changed, and thus DELETES ALL LINE ITEMS
ASSOCIATED TO THE MEMBERSHIP!

The thing is, the line item should not have been allowed to be created in the
first place, as there is a validation in the Manage Future Installments form
that checks if there are pending contributions in the future to add and pay
for the new membership. This validation was failing because it was erronously
checking the contributions' end_date (which doesn't even exist), instead of
the receive_date.

Fixed by using receive_date to validate if there are any pending contributions
in the future.