BASW-415: Fix Membership Line Item Creation on Renewal
When renewing, line items were being created for new memberships, however, the
method that was supposed to return the membership's ID failed to do so,
causing line items to have an empty entity_id field.

Fixed by returning the ID for the new membership as set by the API create
membership call.