BASW-415: Set End Dates on All Line Items When Renewing Plan
Single installment payment plans renewal was only setting end-dates for line
items set to auto-renew. This caused lines not set to renew to neer have an
end date and thus, to always be part of the payment plan.

Fixed by setting end date to all line items in the payment plan, and only
renewing those set to renew.