BASW-220: Use Adjusted Amount as Amount Without Tax
Adjusted amount for first contribution was being taken as the total line item
amount, calculating tax amount from it and the rate for the financial type.

Fixed by using adjusted amount as unit price, multiplying it by the line item
quantity and adding it to the tax amount to calculate line total.