BASW-220: Fix Contribution Total Amount Based on Line Items
When a membership line item is added to a payment plan, this membership is
added to pending contributions using default memberships price set. This price
set is marked as quick_config. When recording a payment, this flag makes
CiviCRM asssume a contribution only has one line item and uses the first one
it finds to calculate contribution amount and tax.

Fixed by recalculating contribution amount on Contribution's pre hook.