BASW-220: Fix Calculation of Contribution Amount on Creation
When creating contributions, calculation of amount on pre-hook was using empty
ID to look for line items. This caused all line itms to be used to calculate
amounts, causing contributions to be created with huge amounts.

Fixed by only doing calculation on contribution update and if ID is not empty.