BASW-220: Overwrite displayLineItems Flag on ContributionView Form
When contributions were being viewed, CiviCRM was checking the is_quick_config
flag for the first line item's price set to decide if all line items should be
shown or not, hiding them if is_quick_config flag was set to true. Since the
default price set is the one we use to add memberships to payment plans, this
caused our line items to remain hidden on most of the payment plans'
contributions.

Fixed by overwriting the preprocessing of displayLineItems flag on the form,
setting it to true if the contribution belongs to a payment plan with more
thank one line item.