BASW-220: Only Set contrib_recur_id if Payment Plan is Auto-renew
Memberships are determined to be auto-renewed or not by checking if they are
related to a recurring contribution via the contrib_recur_id field on the
membership entity. Thus, this field should only be set if the payment plan is
set to auto-renew. This also has implications when editing a payment plan and
changing the auto-renew flag, as setting it should update all memberships
related to the plan to reference the recurring contribution, and unsetting it
should do the opposite.

Fixed by updating the field when editing a recurring contribution and setting
the recur_contrib_id value on the membership only if the plan is auto-renew,
both when updating a reurring contribution and adding a new membership line.