BASW-414: Prevent memebership periods from being extended on pending payment plan payments renewal.
We initialy were extending memeberships at renewal in case they were
renewed with pending payment plan. But the behaviour is diffrenet when
renewing with single contribution. so to ensure consistency we decided
to stop extending the membership at renewal and now it will only get
extended upon completing the first payment.