RSE-67: Fix Transactions on Auto-renewal Jobs
CiviCRM Transactions were using a singleton to create transactions. In
practice, this meant all subsequent transaction for each renewal was in fact
being added to the same transaction. If an exception was thrown on at least on
one of the renewals, everything got rolled back.

Final commit of the transaction is actually done when the transaction object
was destroyed. Added call to object's __destruct(); method to force commit
after each renewal.