CGP-1450: Fix Renewal of Old Plans With Multiple Installments
When upgrading, a new custom group, payment_plan_periods, is added to the
recurring contribution entity. For existing payment plans, records for this
new custom group are created, and initialized by setting the 'next_period' and
'prev_period' fields to 0. This caused a problem when upgrading old plans, as
only plans with next_perio = NULL where being picked for auto-renew.