BASW-222: Fixing an issue with the ability to add future installments
-- Problem

Users can no longer add future installments inside the manage Next
period screen, and an error in the browser console appears that says :

Invalid financial type id passed

The error is caused by getMembershipType() function inside
NextPeriodLineItemHandler.js file which use JQuery grep() method
to find the matching membership type from nextPeriodMembershipTypes
dictionary which contain the list of membership types for next period,
but grep works with arrays so the function end up returing an empty
dictionary which result in not being able to fetch the financial type
for the selected memebership type and thus the error appear.

-- Solution

I replaced the the call to JQuery grep() method with the expression (key
in Dic) to check for the exsitance of the membership type, which fixes
the issue.