MAE-169: Use New Plan Lines to Renew Memberships
When renewing a payment plan, the line items that were being used to renew or
create memberships was the old recurring contribution, instead of the new one.
This meant the new line items were not being related to the memberships, while
the lines of the old plan were.

Fixed by specifying which payment plan was to be used to renew and create
missing memberships.