MAE-259: Remove Deleted Lines from Installments
When deleting lines from a payment plan, respective lines on contributions
on installments were being set to 0, but still showed on contributions and
invoices, which was awkward for users.

Fixed by doing a soft delete of the line after financial calculations are
done, by removing it from the related contribution, setting contribution_id to
null on the line.