MAE-271: Fix Removal of Multiple Donation Items from a Plan
Removal would only work if there was a single line that was not a membership,
as when there were multiple lines, the system did not have a reliable way to
identify the line that should be removed.

Fixed by adding more details to identify the line, and removing the first line
found if several meet the criteria, since as long as the financial type and
amounts match, it doesn't matter wht line is removed.