MAE-477: Remove payment plan updater step
The payment plan updater step that creates subscription line items
and some other info and fields related to payment plans is removed.

The reason is such data is complex and the updater makes many
assumptions where rather such kind of data should be created
as part of data import process.