MAE-477: Temporarily fix for failed renewal tests
The tests updated in this commit will fail we run them on or after March in a
year that comes after a leap year (for example if we run them in march
or april 2021, since 2020 is a leap year).

The reason can be explained in this example :

- Today is 3/3/2021 , these tests will create a payment plan that starts
two years ago (3/3/2019), and the cycle day for this payment plan will
be 61.

- When we renew the payment plan, it will calculate the new
contribution receive based on various factors, but in this case it
should be same day and month as last year which is 3/3/2020 .

- Though before saving the new contribution receive date, we rectify
the amount to be aligned with the payment plan cycle day (because
we allow users to change cycle day if they want to).

- But 2020 is a leap year, and the day 61 would be 2/3/2020 not
3/3/2020, and thus the tests that expect it to be 3/3/2020 will fail.

In this commit I did not resolve this issue, we have a ticket for that
already to resolve the leap year problem wherever it exsist in
Membershipextras, so here I am just temorarily mitigating the problem
by changing the tests to create payment plan in 3 years away from now
to avoid the leap year, which means the issue will happen again in 2022,
but we should fix the issue before then.