MAE-499: Fix confirm dialog button
CiviCRM removes submitOnce function from the core and the confirmation dialog relies on the function. Since CiviCRM does not provide js function as well as provide the API or function for confirmation dialog when submitting the form, we need to implement this functionality for this page.

This commit also refactors the code to improve the code structure, readability of the code.