MAE-522: Prevent extending membership end date
Usually recording the payment for the membership's contribution means :
- Activating the membership by changing its status from pending to
active.
- Renewing the membership by extending its end_date.

At Compucorp, we use the "biz.jmaconsulting.lineitemedit" extension to
add and edit the line items.

Adding a line item changes the contribution status to "Partially Paid"
and recording a payment for an active membership's contribution will
renew it by default and extend its end_date, but this is undesirable
behaviour and this commit will prevent it.