MAE-581: Prevent updating 'Next contribution date' from outside Membershipextras
I here prevent any update to the recurring contribution 'next_sched_contribution_date' field from outside
Membershipextras (Basicly preventing CiviCRM core and the webform_civicrm module from altering the value).

I achive that by removing the value of this field inside the Recurring Contribution Pre hook, this means that
any API or DAO that try to change the value of this field will not work (which is how CiviCRM core and webform_civicrm module alter its value), Which allow us to control its value within
Membershipextras using Direct SQL queieres.

We do that because this field will be used to control autorenewal, so we don't want its value to be altered form places that
we don't control.