MAE-591: Fix renewal job to use recur contribution statuses
The autorenewal job excludes cancelled or refunded recurring contributions from renewal,
but there was two issues with this:

1- We were using the contribution status option group instead of the recurring contribution status option group, it is just happened
that the Values for the option values in each option group are almost simialr such as for the  "Cancelled" status were both equal 3.

2- The refunded status does not exist for recurring contributions, so I removed it from the query that get the recurring contributions to renew and just
kept the part where it check for non cancelled recurring contributions.