PROD-15: Remove code that depends on removed CiviCRM core patch
Here I removed code that depends on removed CiviCRM core patch the introduce hook_civicrm_preSave hook, which
we implemented here to prevent CiviCRM from canceling the membership if any of the payment plan installments
got cancelled.

The reason for that is that we now rely on disabling a core extension called 'contributioncancelactions'
which should give us the behaviour we trying to accomplish.