MAE-612: Change the autorenewal job trigger back to use membership end date instead of next scheduled contribution date
There was a misunderstanding that the trigger for autorenewal should be the next scheduled contribution date. There is a risk that this date may be before or
after the end date of the membership (specially in Direct debit payment plans).

As such here we change the "trigger" for the autorenwal job back so it uses the end date of the membership (taking into account the payment plan setting Days to renew in advance),
or the minimum membership end date if there are more than one membership in the payment plan with different end dates.

The installments receive dates is still getting calculated based on the next scheduled contribution date, so nothing is changing in that regard.