MAE-376: Fix failed test
The fixed test relied on dynamic calculation for the expected value, which resulted in the test to fail at certain times of the year. for example if you run the test
at 2021-11-25, then the expected end date will be 2022-09-30 but the test assumes that it would be 2021-09-30 (see $expectedEndDate = new DateTime(date('Y-09-30')), it uses current year).

So here I removed the dynamic calculation and relied on specific hardcoded expected value and specific start date.