MAE-634: Remove obsolete code from BASW-175 work
BASW-175 that allowed users to be able to adjust Payment plans total amount manually, but on Membership V5 we simplified many things and one of these simplification was to prevent users from manually altering the
payment plan amount (and instead they can use price sets/fields to give memberships to members using non
default prices). This commit removes the code that was responsible for recalculating the tax amount and line items
amounts if the user change the amount manually since it is no longer relevant.