HCSASPRT-66: Fix Manage Instalments page issues when number of instalments is NULL
When the reccuring contribution number of instalments field is NULL, it results in not being able
to load the manage instalments page, given that in that page controler we default the instalments count
to 0 if its value is NULL, which we later use as a divisor in one of the calculations in that page.

Here I am chaning the default value for the instalments field to 1 instead of 0 in case its value is NULL,
not just in the affected page but also in other places that are not really affected, but just as precaution.